% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LarsPath.R
\name{listToMatrix}
\alias{listToMatrix}
\title{List to sparse matrix conversion}
\usage{
listToMatrix(x, row = c("covariates", "lambda"))
}
\arguments{
\item{x}{a \code{\link{LarsPath}} or \code{EMlasso} object}

\item{row}{if covariates, covariates are in row}
}
\value{
A sparse matrix containing the values of estimated coefficients for all penalty parameter and all covariates
}
\description{
create a matrix with all estimated coefficients from the output of \code{\link{HDlars}} or \code{\link{EMlasso}} functions.
}
\seealso{
\code{\link{HDlars}} \code{\link{EMlasso}}
}
