% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LarsPath.R
\docType{methods}
\name{plot-methods}
\alias{plot-methods}
\alias{plot,LarsPath-method}
\title{plot methods for LarsPath object}
\usage{
\S4method{plot}{LarsPath}(
  x,
  sep.line = FALSE,
  abscissa = c("l1norm", "lambda"),
  log.scale = FALSE,
  ...
)
}
\arguments{
\item{x}{LarsPath object}

\item{sep.line}{If TRUE, print vertical dashed line when a variable is added or dropped in the path}

\item{abscissa}{either "l1norm" or "lambda". If "lambda", regularization parameter is used as abscissa, else l1 norm of the solution is used.}

\item{log.scale}{If TRUE, use logarithm scale on abscissa}

\item{...}{Other plot arguments}
}
\description{
plot the path of the lars algorithm.
}
\seealso{
\code{\link{HDlars}} \code{\link{LarsPath}}
}
