% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvlars.R
\name{plot.HDcvlars}
\alias{plot.HDcvlars}
\title{plot cross validation mean square error}
\usage{
\method{plot}{HDcvlars}(x, ...)
}
\arguments{
\item{x}{Output from HDcvlars function.}

\item{...}{graphical parameters}
}
\description{
plot cross validation mean square error
}
\examples{
dataset <- simul(50, 10000, 0.4, 10, 50, matrix(c(0.1, 0.8, 0.02, 0.02), nrow = 2))
result <- HDcvlars(dataset$data, dataset$response, 5)
plot(result)
}
\author{
Quentin Grimonprez
}
