% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LarsPath.R
\name{plotCoefficient}
\alias{plotCoefficient}
\title{Plot of coefficients}
\usage{
plotCoefficient(x, step, ylab = "coefficients", xlab = "variables", ...)
}
\arguments{
\item{x}{A LarsPath object.}

\item{step}{The step at which you want to plot the coefficients.}

\item{ylab}{Name of the y axis.}

\item{xlab}{Name of the x axis.}

\item{...}{Other plot arguments.}
}
\description{
Plot of the coefficients of a step
}
\examples{
dataset <- simul(50, 1000, 0.4, 10, 50, matrix(c(0.1, 0.8, 0.02, 0.02), nrow = 2))
result <- HDfusion(dataset$data, dataset$response)
plotCoefficient(result, result@nbStep) # plot coefficients at the last step
}
\seealso{
\code{\link{HDlars}} \code{\link{LarsPath}}
}
