% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCv}
\alias{plotCv}
\title{plot cross validation mean square error}
\usage{
plotCv(x)
}
\arguments{
\item{x}{Output from HDcvlars function.}
}
\description{
plot cross validation mean square error
}
\examples{
dataset=simul(50,10000,0.4,10,50,matrix(c(0.1,0.8,0.02,0.02),nrow=2))
result=HDcvlars(dataset$data,dataset$response,5)
plotCv(result)
}
\author{
Quentin Grimonprez
}

