% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDPenReg-package.R
\docType{package}
\name{HDPenReg-package}
\alias{HDPenReg-package}
\alias{HDPenReg-package,}
\alias{HDPenReg}
\title{Algorithms for lasso and fused-lasso problems.}
\description{
This package contains algorithms for lasso and fused-lasso problems.
It contains an implementation of the lars algorithm [1],
for the lasso and fusion penalization and EM-based algorithms for (logistic) lasso and fused-lasso.
}
\details{
\tabular{ll}{
Package: \tab HDPenReg\cr
Type: \tab Package\cr
Version: \tab 0.94.5\cr
Date: \tab 2019-03-29\cr
License: \tab GPL (>=2) \cr
}

The main function is \link{HDlars}.
}
\examples{
\dontrun{
# see vignette
vignette("HDPenReg")
}

}
\seealso{
\code{\link{HDlars}} \code{\link{HDcvlars}}
}
\author{
Maintainer: Quentin Grimonprez <quentin.grimonprez@inria.fr>
}
\keyword{package}
