% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{HDPenReg-package}
\alias{HDPenReg}
\alias{HDPenReg-package}
\alias{HDPenReg-package,}
\title{Algorithms for lasso and fused-lasso problems.}
\description{
This package contains algorithms for lasso and fused-lasso problems.
It contains an implementation of the lars algorithm [1],
for the lasso and fusion penalization and EM-based algorithms for (logistic) lasso and fused-lasso.
}
\details{
\tabular{ll}{
Package: \tab HDPenReg\cr
Type: \tab Package\cr
Version: \tab 0.90\cr
Date: \tab 2014-10-23\cr
License: \tab GPL (>=2) \cr
}

The main function is \link{HDlars}.
}
\examples{
#see vignette
#vignette("HDPenReg")
}
\author{
Maintainer: Quentin Grimonprez <quentin.grimonprez@inria.fr>
}
\keyword{package}

