% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{HDPenReg-package}
\alias{HDPenReg}
\alias{HDPenReg-package}
\alias{HDPenReg-package,}
\title{Algorithms for lasso and fused-lasso problems.}
\description{
This package contains an implementation of the lars algorithm [1],
for the lasso and fusion penalization. It works even if the number of
covariate is greater than the number of individuals.
}
\details{
\tabular{ll}{
Package: \tab HDPenReg\cr
Type: \tab Package\cr
Version: \tab 0.89.1\cr
Date: \tab 2014-05-28\cr
License: \tab GPL (>=2) \cr
}

The main function is \link{HDlars}.
}
\examples{
#see vignette
#vignette("HDPenReg")
}
\author{
Maintainer: Quentin Grimonprez <quentin.grimonprez@inria.fr>
}
\keyword{package}

