\docType{methods}
\name{plot-methods}
\alias{plot,LarsPath-method}
\alias{plot-methods}
\title{plot methods for LarsPath object}
\description{
  plot the path of the lars algorithm.
}
\details{
  \describe{ \item{x}{LarsPath object.} \item{sep.line}{If
  TRUE, print vertical dashed line when a variable is added
  or dropped in the path.} \item{...}{Other plot
  arguments.} }
}

