% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LarsPath.R
\docType{methods}
\name{plot-methods}
\alias{plot,LarsPath-method}
\alias{plot-methods}
\title{plot methods for LarsPath object}
\usage{
\S4method{plot}{LarsPath}(x, sep.line = TRUE, ...)
}
\arguments{
\item{x}{LarsPath object}

\item{sep.line}{If TRUE, print vertical dashed line when a variable is added or dropped in the path}

\item{...}{Other plot arguments}
}
\description{
plot the path of the lars algorithm.
}

